inflearn-llm-app/pyproject.toml
[project]
name = "inflearn-llm-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt>=0.9",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.0",
    "langchainhub>=0.1.21",
    "openai>=2.14.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

